# examples/cactus-example-cbdc-bridging-backend/src/main/go/generated/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    name: Hyperledger Cacti
  description: |
    Cactus-Example
  title: CBDC-example backend API
  version: 0.0.2
  x-logo:
    altText: Cacti logo
    backgroundColor: '#fafafa'
    url: https://github.com/hyperledger/cacti/blob/main/images/HL_Cacti_Logo_Color.png
servers:
- url: /
...
paths:
  /api/v1/@hyperledger/cactus-example-cbdc/transact:
    post:
      description: Allows users to queue intents for transactions based on specified
        parameters.
      operationId: Transact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactResponse'
          description: Transaction successfully queued
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: An error occurred
      summary: Submit a transaction intent
      tags:
      - transact
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-example-cbdc/transact
...
components:
  schemas:
    TransactRequest:
      description: "Request schema for initiating a transaction. Includes details\
        \ such as the transaction context, mode (data or transfer), payload, and information\
        \ about the source and destination DLT networks."
      example:
        sourceChain:
          assetType: FABRIC
        amount: "100"
        receiver: user2
        sender: user1
        receiverChain:
          assetType: FABRIC
      properties:
        sender:
          example: user1
          type: string
        receiver:
          example: user2
          type: string
        sourceChain:
          $ref: '#/components/schemas/AssetType'
        receiverChain:
          $ref: '#/components/schemas/AssetType'
        amount:
          example: "100"
          type: string
      required:
      - amount
      - receiver
      - receiverChain
      - sender
      - sourceChain
      type: object
    AssetType:
      description: Enum for the type of asset being transferred.
      example:
        assetType: FABRIC
      properties:
        assetType:
          enum:
          - FABRIC
          - BESU
          type: string
      type: object
    TransactResponse:
      description: Response schema for a transaction request.
      example:
        statusResponse:
          sessionID: 123e4567-e89b-12d3-a456-426614174000
          status: NOT_FOUND
      properties:
        statusResponse:
          $ref: '#/components/schemas/StatusResponse'
      required:
      - statusResponse
      type: object
    StatusResponse:
      description: Response schema for a transaction status request.
      example:
        sessionID: 123e4567-e89b-12d3-a456-426614174000
        status: NOT_FOUND
      properties:
        sessionID:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        status:
          enum:
          - NOT_FOUND
          - INVALID
          - PENDING
          - DONE
          - FAILED
          type: string
      required:
      - status
      type: object
    APIError:
      description: An Error
      example:
        code: 16384
        type: bad-request
        message: Oops there's been an internal error handling
        status: 500
        timestamp: 2021-07-21T17:32:28Z
      properties:
        type:
          description: HTTP error type
          example: bad-request
          type: string
        code:
          description: Numeric error code
          example: 16384
          type: integer
        status:
          description: HTTP status of the error
          example: 500
          type: integer
        message:
          description: Long error description
          example: Oops there's been an internal error handling
          type: string
        timestamp:
          description: Timestamp of the error
          example: 2021-07-21T17:32:28Z
          type: string
      required:
      - code
      - message
      - status
      - timestamp
      - type
      type: object
      x-category: response
      x-go-type: ApiError
      x-go-name: ApiError
...